+++ mingling/src/example_docs.rs
@@ -128,8 +128,7 @@ pub mod example_argument_parse {}
 ///
 ///  However, you will lose some capabilities:
 ///
-///  1. `&mut` resource injection is not available in async chain functions
-///  2. The program will not be able to use panic unwind functionality
+///  1. The program will not be able to use panic unwind functionality
 ///
 ///  Run:
 ///  ```bash